requirements.txt
asgiref==3.6.0
Django==4.1.6
django-jazzmin==2.6.0
django-ninja==0.20.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
idna==3.4
Jinja2==3.1.2
jose==1.0.0
MarkupSafe==2.1.2
pip-api==0.0.30
pyasn1==0.4.8
pydantic==1.10.4
python-jose==3.3.0
rsa==4.9
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
utils==1.0.1
